Rename purple.h to quiet cppcheck
